refactor(pdf-service): move /app/domains/share/services/pdf/pdfStyles.ts to /app/services/pdf/
