+++ app/services/pdf/attachment/createAttachmentEntries.ts
@@ -1,6 +1,6 @@
 import type PDFDocument from "pdfkit";
 import type { AttachmentEntries } from "~/services/pdf/attachment";
-import { pdfStyles } from "../../../domains/shared/services/pdf/pdfStyles";
+import { pdfStyles } from "../pdfStyles";
 
 export function createAttachmentEntries(
   doc: typeof PDFDocument,
